fixing a lil bugs 5

+++ src/crypto-algorithms/cuda-version/arcfour.cu
@@ -92,7 +92,7 @@ void enc_file(char *filename, char *enc_filename)
     size_t len;
     BYTE *d_data = NULL;
     BYTE *d_key = NULL;
-    size_t *d_len = NULL;
+    int *d_len = NULL;
     cudaError_t err = cudaSuccess;
     
     data = read_file(filename);
@@ -103,7 +103,7 @@ void enc_file(char *filename, char *enc_filename)
     err = cudaMalloc(&d_data, len * sizeof(BYTE));
     print_error_message(err, (const char *) "d_data", ALLOC); 
 
-    err = cudaMalloc(&d_len, sizeof(size_t));
+    err = cudaMalloc(&d_len, sizeof(int));
     print_error_message(err, (const char *) "d_len", ALLOC);
 
     err = cudaMalloc(&d_key, 1024 * sizeof(BYTE));
@@ -112,7 +112,7 @@ void enc_file(char *filename, char *enc_filename)
     err = cudaMemcpy(d_data, data, len * sizeof(BYTE), cudaMemcpyHostToDevice);
     print_error_message(err, (const char *) "d_data", COPY);
 
-    err = cudaMemcpy(d_len, &len, sizeof(size_t), cudaMemcpyHostToDevice);
+    err = cudaMemcpy(d_len, &len, sizeof(int), cudaMemcpyHostToDevice);
     print_error_message(err, (const char *) "d_len", COPY);
 
     err = cudaMemcpy(d_key, 1024, sizeof(size_t), cudaMemcpyHostToDevice);
